% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\alias{Model}
\title{A reference class that provides a template for models}
\description{
This class only provides the basic elements of a model. Subclasses should be
created to accommodate different situations and goals.
}
\section{Fields}{

\describe{
\item{\code{paramlist}}{a vector of parameter names}

\item{\code{nparam}}{the number of parameters}

\item{\code{startparam}}{a vector of starting parameter values for estimation}

\item{\code{indices}}{a list giving the index of each parameter}

\item{\code{lhslist}}{names of dependent (i.e., stochastic) variables}

\item{\code{rhslist}}{names of independent (i.e., non-stochastic) variables}
}}

\section{Methods}{

\describe{
\item{\code{isValidModel()}}{A method to determine if it's a valid instantiation and should be overridden in subclasses.}

\item{\code{XBeta(varlist, data, param, include_constant = 0, coef_prefix = "",
  datavar_suffix = "", constname = "constant")}}{Return X * beta}
}}

