% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelData.R
\docType{class}
\name{ModelData-class}
\alias{ModelData-class}
\alias{ModelData}
\title{A Reference Class that Provides a Template for Data}
\description{
A ModelData object stores dataset in the field var, an R data.frame object.
The constructor \code{ModelData()} accepts the same inputs as the constructor for a data frame,
plus an option of specifying variable names and an option of adding a customer list of
constants characterizing the data.
}
\section{Fields}{

\describe{
\item{\code{var}}{A data.frame object to hold dataset}

\item{\code{varnames}}{A vector of variable names in the dataset}

\item{\code{nvars}}{The number of variables in the dataset}

\item{\code{nobs}}{The number of observations in the dataset}

\item{\code{const}}{A list to hold constants characterizing the dataset}

\item{\code{groupvar}}{A vector identifying groups in the panel}

\item{\code{ngroup}}{The number of groups in the panel}

\item{\code{group_size}}{A vector of size ngroups giving the number of observations in each group}

\item{\code{unique_group_sizes}}{Unique group sizes that appear in the dataset}
}}

\section{Methods}{

\describe{
\item{\code{addArrayVars(data, name = "arrayvar")}}{Add a matrix or 3-dimension variable in the field \code{var}.

        \code{data}: A matrix or a 3-dimension array.

        \code{name}: The name of the array variable added.}

\item{\code{addData(..., names = NULL)}}{Add variables to columns.

        \code{...}: The same as the constructor of data.frame.

        \code{names}: Variable names}

\item{\code{collapseArrayVars()}}{Collapse all variables with the name \code{varname_array_d1_d2} to an array variable 
         \code{varname} with dimensions \code{R} = \code{max(d1)} by \code{C = max(d2)}.

         The reverse function of \code{expandArrayVars}.}

\item{\code{expandArrayVars()}}{Expand all array variables \code{varname} with dimensions \code{R} by \code{C}
         into multiple variables with the variable name \code{varname_array_[1:R]_[1:C]}.

         The reverse function of \code{collapseArrayVars}.}

\item{\code{isVariable(varname)}}{Determine if variables are in the dataset.

        \code{varname}: The name of the variables.}

\item{\code{removeData(col)}}{Remove variables from columns.

        \code{col}: Column index or variable names}

\item{\code{selectData(row = 1:.self$nobs, col = .self$varnames)}}{Select a subset of data in the same way as subsetting a data.frame object.

        \code{row}: Index or conditions for selecting rows. Default is all rows.

        \code{col}: Index or variable names for selecting columns. Default is all columns.}

\item{\code{setGroup(value)}}{Add a vector identifying groups in the panel. This vector must be sorted in ascending
        order with length equal to the number of observations in the dataset.

        In most cases, the group variable is also in the data. It's better to first sort the data
        by the group variable, then create the ModelData object and set the group variable.

        \code{value}: A vector identifying groups.}
}}

\examples{
x <- rnorm(5)
y <- rnorm(5)
rhs <- list(x = x, y = y)
strvar <- c("f", "m", "m", "f", "m")
ModelData(x, y, const = list(mu = 0, sigma = 1))
ModelData(rhs, strvar, varnames = c("rhs1", "rhs2", "gender"))
}
