% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToDisk.R
\name{saveToDisk}
\alias{saveToDisk}
\title{Save a \code{ModelData} object to files}
\usage{
saveToDisk(obj, directory, name, precision = 8)
}
\arguments{
\item{obj}{A ModelData object.}

\item{directory}{The location of files to be saved.}

\item{name}{The name of files to be saved.}

\item{precision}{The number of decimal to store for all inputs. Default is 8.}
}
\description{
The field \code{var} is saved as \code{.csv} file and the remaining fields are saved
as a \code{.rds} file. The group variable is added in the \code{var} with name \code{group_export}.
All array variables are expanded.
}
