% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromDisk.R
\name{loadFromDisk}
\alias{loadFromDisk}
\title{Load a \code{ModelData} object from files saved using \code{saveToDisk}}
\usage{
loadFromDisk(obj, directory, name, collapseArrayVars = 1)
}
\arguments{
\item{obj}{A ModelData object, typically an empty one.}

\item{directory}{location of files to be loaded.}

\item{name}{name of files to be loaded.}

\item{collapseArrayVars}{whether array variables are collapsed. Default is collapse.}
}
\description{
Data are read from a \code{.csv} file and other fields are read from a \code{.rds}
file. The group variable with name \code{group_export} is deleted.
}
