% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\alias{Model}
\title{A Reference Class that Provides a Template for Models}
\description{
This class only provides the basic elements of a model. Subclasses should be
created to accommodate different situations and goals.
}
\section{Fields}{

\describe{
\item{\code{paramlist}}{A vector of parameter names.}

\item{\code{nparam}}{The number of parameters.}

\item{\code{startparam}}{A vector of starting parameter values for estimation.}

\item{\code{indices}}{A list giving the index of each parameter.}

\item{\code{lhslist}}{The names of dependent (i.e., stochastic) variables.}

\item{\code{rhslist}}{The names of independent (i.e., non-stochastic) variables.}
}}

\section{Methods}{

\describe{
\item{\code{isValidModel()}}{A method to determine if it's a valid instantiation and should be overridden in subclasses.}

\item{\code{isValidParameterVector(param)}}{A method to determine if the parameters are valid for the model.

        \code{param}: A vector of parameters.}

\item{\code{XBeta(varlist, data, param, constant = 0, constname = "constant",
  coeff_prefix = "", coeff_suffix = "_coeff", datavar_prefix = "",
  datavar_suffix = "")}}{Return the product of data and coefficients.

        \code{varlist}: A vector of variables names.

        \code{data}: A ModelData object.

        \code{param}: A vector of coefficients. The order of the coefficients should be the same as 
        the field \code{varnames} of the \code{Model} object.

        \code{constant}: Whether the calculation includes constant or not.

        \code{constname}: The name of the constant.

        \code{coeff_prefix}: The prefix of coefficient names in the model.

        \code{coeff_suffix}: The suffix of coefficient names in the model.

        \code{datavar_prefix}: The prefix of variable names in the dataset.

        \code{datavar_suffix}: The suffix of variable names in the dataset.}
}}

