% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEstimationOptions.R
\docType{class}
\name{ModelEstimationOptions-class}
\alias{ModelEstimationOptions-class}
\alias{ModelEstimationOptions}
\title{A Reference Class that Defines Options for the \code{Estimate} Method of \code{Model}}
\description{
A Reference Class that Defines Options for the \code{Estimate} Method of \code{Model}
}
\section{Fields}{

\describe{
\item{\code{startparam}}{Starting paramter vector.}

\item{\code{hesstol}}{Numerical step for computing the Hessian.}
}}


