% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEstimationOutput.R
\docType{class}
\name{ModelEstimationOutput-class}
\alias{ModelEstimationOutput-class}
\alias{ModelEstimationOutput}
\title{A Reference Class that Provides a Template for Model Estimation Output.}
\description{
This class only provides the basic elements of the model estimation output.
Subclasses should be created according to different models.
}
\section{Fields}{

\describe{
\item{\code{param}}{Estimated parameters}

\item{\code{value}}{The value of objective function at estimated parameters}

\item{\code{convergence}}{flag for convergence}

\item{\code{model}}{model object used for estimation}

\item{\code{nobs}}{number of observations in the data used for estimation}

\item{\code{vcov}}{Variance-covariance matrix of parameters}

\item{\code{se}}{Standard errors of parameters}
}}


