% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEstimationOutput.R
\docType{class}
\name{ModelEstimationOutput-class}
\alias{ModelEstimationOutput-class}
\alias{ModelEstimationOutput}
\title{A Reference Class that Provides a Template for Model Estimation Output}
\description{
This class only provides the basic elements of the model estimation output.
Subclasses should be created according to different models.
}
\section{Fields}{

\describe{
\item{\code{param}}{Locally optimal solution.}

\item{\code{fval}}{The value of objective function at the optimal solution.}

\item{\code{convergence}}{An integer indicating convergence of estimation. 0 indicates successful completion.}

\item{\code{message}}{A character string giving any additional information returned by the optimizer.}

\item{\code{estopts}}{A \code{ModelEstimationOptions} object.}

\item{\code{model}}{A \code{Model} object used for estimation.}

\item{\code{nobs}}{The number of observations in the data used for estimation.}

\item{\code{vcov}}{Variance-covariance matrix of estimated parameters.}

\item{\code{se}}{Standard errors of estimated parameters.}
}}


